requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx[http2]==0.25.1
parsel==1.8.1
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0